+++ tfhe/src/high_level_api/xof_key_set/internal.rs
@@ -36,12 +36,14 @@ use crate::{
 
 use crate::core_crypto::commons::ciphertext_modulus::CiphertextModulus as CoreCiphertextModulus;
 use crate::high_level_api::keys::expanded::{
-    ExpandedAtomicPatternNoiseSquashingKey, ExpandedAtomicPatternServerKey,
-    ExpandedDecompressionKey, ExpandedKS32AtomicPatternServerKey, ExpandedNoiseSquashingKey,
-    ExpandedStandardAtomicPatternServerKey, IntegerExpandedServerKey,
-    ShortintExpandedBootstrappingKey, ShortintExpandedServerKey,
+    ExpandedAtomicPatternNoiseSquashingKey, ExpandedDecompressionKey, ExpandedNoiseSquashingKey,
+    IntegerExpandedServerKey, ShortintExpandedBootstrappingKey, ShortintExpandedServerKey,
 };
 use crate::high_level_api::keys::{CompactPrivateKey, ReRandomizationKeyGenerationInfo};
+use crate::shortint::atomic_pattern::expanded::{
+    ExpandedAtomicPatternServerKey, ExpandedKS32AtomicPatternServerKey,
+    ExpandedStandardAtomicPatternServerKey,
+};
 use crate::shortint::key_switching_key::KeySwitchingKeyDestinationAtomicPattern;
 use crate::shortint::noise_squashing::atomic_pattern::compressed::standard::CompressedStandardAtomicPatternNoiseSquashingKey;
 use crate::shortint::parameters::ModulusSwitchNoiseReductionParams;
