+++ tfhe/src/shortint/noise_squashing/atomic_pattern/expanded.rs
@@ -1,0 +1,6 @@
+use crate::shortint::server_key::expanded::ShortintExpandedBootstrappingKey;
+
+pub enum ExpandedAtomicPatternNoiseSquashingKey {
+    Standard(ShortintExpandedBootstrappingKey<u128, u64>),
+    KeySwitch32(ShortintExpandedBootstrappingKey<u128, u32>),
+}
